property refactoring (#473)
