+++ src/components/bundles/bundles_list.tsx
@@ -2,6 +2,12 @@ import { useMediaQuery, useTheme } from "@mui/material";
 import { BackendApi } from "../../backend/backend_api";
 import BundlesListDesktop from "./bundles_list_desktop";
 import BundlesListMobile from "./bundles_list_mobile";
+import { useEffect } from "react";
+import { BundleData } from "../../backend/bundle_data";
+import { startLoading, finishLoading, addBundle, setMaxActiveBundles, reset } from "../../redux/slices/bundles";
+import address from "../address";
+import { useDispatch, useSelector } from "react-redux";
+import { RootState } from "../../redux/store";
 
 export interface BundlesProps {
     backend: BackendApi;
@@ -10,10 +16,36 @@ export interface BundlesProps {
 export default function BundlesList(props: BundlesProps) {
     const theme = useTheme();
     const isMobile = useMediaQuery(theme.breakpoints.down('md'));
+    const dispatch = useDispatch();
+    const bundleManagementApi = props.backend.bundleManagement;
+    const signer = useSelector((state: RootState) => state.chain.signer);
+
+    useEffect(() => {
+        async function getBundles() {
+            if (signer === undefined) {
+                return;
+            }
+
+            dispatch(startLoading());
+            dispatch(reset());
+            
+            if (address === undefined ) {
+                dispatch(finishLoading());
+                return;
+            }
     
+            await bundleManagementApi.fetchAllBundles((bundle: BundleData) => dispatch(addBundle(bundle)) );
+            const maxActiveBundles = await bundleManagementApi.maxBundles();
+            dispatch(setMaxActiveBundles(maxActiveBundles));
+            dispatch(finishLoading());
+        }
+        getBundles();
+    }, [signer, bundleManagementApi, dispatch]); // update bundles when signer changes
+
+
     if (isMobile) {
-        return (<BundlesListMobile backend={props.backend} />);
+        return (<BundlesListMobile />);
     } else {
-        return (<BundlesListDesktop backend={props.backend} />);
+        return (<BundlesListDesktop usd2={props.backend.usd2} usd2Decimals={props.backend.usd2Decimals} />);
     }
 }
